Rewrite asset handling to use diff middleware based on env
